Vermeidung NPE

+++ jverein/src/de/jost_net/JVerein/gui/control/MitgliedControl.java
@@ -1,7 +1,7 @@
 /**********************************************************************
  * $Source: /home/xubuntu/berlios_backup/github/tmp-cvs/jverein/Repository/jverein/src/de/jost_net/JVerein/gui/control/MitgliedControl.java,v $
- * $Revision: 1.87 $
- * $Date: 2010/09/03 16:52:42 $
+ * $Revision: 1.88 $
+ * $Date: 2010/09/06 17:46:14 $
  * $Author: jost $
  *
  * Copyright (c) by Heiner Jostkleigrewe
@@ -9,7 +9,10 @@
  * heiner@jverein.de
  * www.jverein.de
  * $Log: MitgliedControl.java,v $
- * Revision 1.87  2010/09/03 16:52:42  jost
+ * Revision 1.88  2010/09/06 17:46:14  jost
+ * Vermeidung NPE
+ *
+ * Revision 1.87  2010-09-03 16:52:42  jost
  * Mitgliedsfoto wird jetzt korrekt skaliert.
  *
  * Revision 1.86  2010-09-01 05:56:57  jost
@@ -2179,37 +2182,40 @@ public class MitgliedControl extends AbstractControl
       }
       m.store();
 
-      Mitgliedfoto f = null;
-      DBIterator it = Einstellungen.getDBService().createList(
-          Mitgliedfoto.class);
-      it.addFilter("mitglied = ?", new Object[] { m.getID() });
-      if (it.size() > 0)
+      if (Einstellungen.getEinstellung().getMitgliedfoto())
       {
-        f = (Mitgliedfoto) it.next();
-        if (foto == null)
+        Mitgliedfoto f = null;
+        DBIterator it = Einstellungen.getDBService().createList(
+            Mitgliedfoto.class);
+        it.addFilter("mitglied = ?", new Object[] { m.getID() });
+        if (it.size() > 0)
         {
-          f.delete();
+          f = (Mitgliedfoto) it.next();
+          if (foto == null)
+          {
+            f.delete();
+          }
+          else
+          {
+            f.setFoto((byte[]) foto.getValue());
+            f.store();
+          }
         }
         else
         {
+          f = (Mitgliedfoto) Einstellungen.getDBService().createObject(
+              Mitgliedfoto.class, null);
+          f.setMitglied(m);
           f.setFoto((byte[]) foto.getValue());
           f.store();
         }
       }
-      else
-      {
-        f = (Mitgliedfoto) Einstellungen.getDBService().createObject(
-            Mitgliedfoto.class, null);
-        f.setMitglied(m);
-        f.setFoto((byte[]) foto.getValue());
-        f.store();
-      }
-
       if (eigenschaftenTree != null)
       {
         if (!getMitglied().isNewObject())
         {
-          it = Einstellungen.getDBService().createList(Eigenschaften.class);
+          DBIterator it = Einstellungen.getDBService().createList(
+              Eigenschaften.class);
           it.addFilter("mitglied = ?", new Object[] { getMitglied().getID() });
           while (it.hasNext())
           {
@@ -2244,7 +2250,8 @@ public class MitgliedControl extends AbstractControl
           it0.addFilter("label = ?", new Object[] { ti.getName() });
           Felddefinition fd = (Felddefinition) it0.next();
           // Ist bereits ein Datensatz für diese Definiton vorhanden ?
-          it = Einstellungen.getDBService().createList(Zusatzfelder.class);
+          DBIterator it = Einstellungen.getDBService().createList(
+              Zusatzfelder.class);
           it.addFilter("mitglied =?", new Object[] { m.getID() });
           it.addFilter("felddefinition=?", new Object[] { fd.getID() });
           Zusatzfelder zf = null;
